Add context switch to task blocking macros
So far, macros used for blocking tasks have put the CPU to sleep,
preventing other tasks that may be ready to run from running, while
there may still be CPU time left in the current tick period.

Addressing issue #1, macros have been added to expand <blockt> and
<blockv> functionality. These macros accept an additional argument so
that the currently running task's bit in the ready to run variable can
be modified.

Since the task switcher tick interrupt performs what we want for doing
a context switch except for decrementing counters, the most significant
bit in the ready to run variable is now used to have the tick interrupt
skip decrementing of counters when yielding.