Change to AVRASM2 preprocessor directives
AVRASM2 preprocessor directives are not backwards compatible, so older
MCUs are no longer supported. Which is not a problem really because
tickbyte is meant mainly for devices supported by AVRASM2.

These directives allow logical operators with conditional assembly of
sleep mode initialization.

Also fixed <and> that should be <or> when TASK_YIELD is called.